requirements.txt
fastapi
uvicorn
passlib[bcrypt]

python-dotenv
PyJWT
